perf: u8-encoded root permutations throughout the twisted-conjugacy path (E7 10.3s/4.1GB -> 8.4s/2.2GB)
weyl_permutations, piece/element root permutations, class keys, and dual
correspondence permutations now store root indices as u8 (roots < 256)
instead of usize: E7's 2.9M x 126 permutation table drops from ~1.7GB to
~210MB. Byte-identical (316 + 230 tests, E6/E7 diff).

+++ crates/atlas-real-group/src/cartan_class.rs
@@ -58,7 +58,7 @@ pub struct TwistedConjugacyPartition {
     datum: BasedRootDatum,
     distinguished: RootInvolutionData,
     classes: Vec<TwistedConjugacyClass>,
-    class_by_permutation: BTreeMap<Vec<usize>, usize>,
+    class_by_permutation: BTreeMap<Vec<u8>, usize>,
 }
 
 impl TwistedConjugacyPartition {
@@ -66,7 +66,7 @@ impl TwistedConjugacyPartition {
         datum: BasedRootDatum,
         distinguished: RootInvolutionData,
         classes: Vec<TwistedConjugacyClass>,
-        class_by_permutation: BTreeMap<Vec<usize>, usize>,
+        class_by_permutation: BTreeMap<Vec<u8>, usize>,
     ) -> Self {
         Self {
             datum,
@@ -84,7 +84,7 @@ impl TwistedConjugacyPartition {
     /// root-image permutation is `permutation`: the map lookup behind
     /// [`Self::class_of`] without the provenance gates, for consumers that
     /// derive the permutation from a lattice involution directly.
-    pub(crate) fn class_index_of_permutation(&self, permutation: &[usize]) -> Option<usize> {
+    pub(crate) fn class_index_of_permutation(&self, permutation: &[u8]) -> Option<usize> {
         self.class_by_permutation.get(permutation).copied()
     }
 
@@ -104,11 +104,11 @@ impl TwistedConjugacyPartition {
         {
             return Err(StructureError::DistinguishedInvolutionMismatch);
         }
-        let key: Vec<usize> = twisted
+        let key: Vec<u8> = twisted
             .root_involution()
             .image_permutation()
             .iter()
-            .map(|id| id.0)
+            .map(|id| id.0 as u8)
             .collect();
         self.class_by_permutation.get(&key).copied().ok_or(
             StructureError::CartanClassificationInvariantViolation {
